refactor: split commands/index.ts into domain-specific files
Break the 825-line command registration monolith into focused modules:
panes.ts, sessions.ts, watches.ts, ui.ts, tasks.ts. index.ts shrinks
to 15 lines that just imports and calls register functions.
